✨ feat(queue): introduce PaperQueueCommon interface and unify queue operations
- add PaperQueueCommon interface for shared queue operations (delete, forceCleanup)
- update PaperQueueImpl and PaperOwnedQueueImpl to implement PaperQueueCommon
- update queue commands to use PaperQueueCommon for type safety and flexibility
- refactor LuckpermsPriorityHook to LuckpermsPriorityResolver and update usage
- update VelocityQueueImpl to extend AbstractTickableQueue and accept scheduler
- fix spinner index calculation in QueueDisplay
- update QueueListCommand to use getAllUuidsOrderedByPosition
- add debug prints to DefaultQueueStartHook and PolarQueueStartHook

+++ surf-queue-velocity/src/main/kotlin/dev/slne/surf/queue/velocity/queue/QueueDisplay.kt
@@ -25,7 +25,7 @@ class QueueDisplay(private val queue: VelocityQueueImpl) {
 
     private suspend fun updateActionBars() {
         val uuidsWithPosition = cachedUuidsWithPosition ?: return
-        val spinnerIndex = (queue.tickCount % spinner.size)
+        val spinnerIndex = (queue.tickCount % spinner.size).toInt()
         val spinnerEnd = spinner[spinnerIndex]
         val spinnerStart = spinnerReversed[spinnerIndex]
         val paused = queue.isPaused()
